Fix RestartForceExitStatus and SuccessExitStatus
RestartForceExitStatus and SuccessExitStatus both require spaces between the exit codes. From http://www.freedesktop.org/software/systemd/man/systemd.service.html:

"Exit status definitions can either be numeric exit codes or termination signal names, separated by spaces. For example:

SuccessExitStatus=1 2 8 SIGKILL

ensures that exit codes 1, 2, 8 and the termination signal SIGKILL are considered clean service terminations."

The existing 133 will only force a restart if the exit code is 133. I was getting exit code 3 and it was not restarting:

● docker-newrelic.service - docker-newrelic
   Loaded: loaded (/etc/systemd/system/docker-newrelic.service; enabled; vendor preset: disabled)
   Active: failed (Result: exit-code) since Fri 2015-04-24 18:06:49 UTC; 25min ago
     Docs: man:systemd-nspawn(1)
  Process: 9392 ExecStart=/usr/bin/systemd-nspawn --quiet --keep-unit --share-system --bind=/dev:/dev --bind=/sys:/sys --bind=/proc:/proc --capability=all --directory=/var/lib/container/docker-newrelic --machine=docker-newrelic /bin/sh -c nrsysmond-config --set license_key=hahanope && nrsysmond -c /etc/nrsysmond.cfg -n %H -d info -l /dev/stdout -f (code=exited, status=3)
  Process: 9387 ExecStartPre=/bin/sh -c cd /var/lib/container/docker-newrelic/ && tar -xf image.tar (code=exited, status=0/SUCCESS)
  Process: 9379 ExecStartPre=/bin/sh -c docker export docker-newrelic > /var/lib/container/docker-newrelic/image.tar (code=exited, status=0/SUCCESS)
  Process: 9362 ExecStartPre=/bin/sh -c docker rm -f docker-newrelic; docker create --name docker-newrelic johanneswuerbach/newrelic-sysmond (code=exited, status=0/SUCCESS)
  Process: 9359 ExecStartPre=/usr/bin/mkdir -p /var/lib/container/docker-newrelic (code=exited, status=0/SUCCESS)
  Process: 9356 ExecStartPre=/usr/bin/rm -rf /var/lib/container/docker-newrelic (code=exited, status=0/SUCCESS)
 Main PID: 9392 (code=exited, status=3)
   Status: "Terminating..."

The alternative would be to set:

Restart=always
RestartSec=60

You could skip RestartSec or change it to a different value based on your needs, the default is 100ms so that's a bit fast if you're having a real problem with the service.